requirements.txt
GitPython==3.1.43
icecream==2.1.3
inquirer==3.4.0
loguru==0.7.2
pyperclip==1.9.0
